refactor: split roadmapview across sites/brc-analytics and sites/ga2 (#1550)
