+++ app/components/content/ga2/index.tsx
@@ -1,2 +1,1 @@
 export { default as SectionAbout } from "./sectionAbout.mdx";
-export { default as SectionRoadmap } from "./sectionRoadmap.mdx";
