+++ sites/brc-analytics/views/RoadmapView/roadmapView.tsx
@@ -1,9 +1,9 @@
-import { SectionRoadmap } from "@/components/content";
 import { SectionHero } from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero";
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import { Typography } from "@mui/material";
 import { Fragment, JSX } from "react";
-import { BREADCRUMBS } from "./common/constants";
+import { BREADCRUMBS } from "./constants";
+import SectionRoadmap from "./content/sectionRoadmap.mdx";
 
 export const RoadmapView = (): JSX.Element => {
   return (
