+++ sites/ga2/views/RoadmapView/constants.ts
@@ -1,0 +1,7 @@
+import { Breadcrumb } from "@databiosphere/findable-ui/lib/components/common/Breadcrumbs/breadcrumbs";
+
+export const BREADCRUMBS: Breadcrumb[] = [
+  { path: "/", text: "Home" },
+  { path: "/about", text: "About" },
+  { path: "", text: "Roadmap" },
+];
