+++ sites/ga2/views/RoadmapView/roadmapView.tsx
@@ -1,7 +1,7 @@
-import { SectionRoadmap } from "@/components/content/ga2";
 import { SectionHero } from "@/components/Layout/components/AppLayout/components/Section/components/SectionHero/sectionHero";
 import { Fragment, JSX } from "react";
-import { BREADCRUMBS } from "../common/constants";
+import { BREADCRUMBS } from "./constants";
+import SectionRoadmap from "./content/sectionRoadmap.mdx";
 
 export const RoadmapView = (): JSX.Element => {
   return (
